Move remaining tests
